Added delete page functionality and more error messages for trying to access pages, admins that don'e exist

+++ delete-admin.php
@@ -1,14 +1,27 @@
 <?php
-session_start();
+require "header.php";
 require "session-auth.php";
 //checks if user is allowed to go here
+echo "<a href='pages.php'>Back to list of pages</a><br>";
+//Link back to pages list if user has error
 $adminId = $_GET["user"];
 require "db.php";
-$sql = "DELETE FROM `cms_users` WHERE `userId` = :adminId";
-$cmd = $db->prepare($sql);
-$cmd->bindParam(":adminId", $adminId, PDO::PARAM_INT);
-$cmd->execute();
-//Deletes the entry with the assocciated id
+$sqlCheck = "SELECT `userId` FROM `cms_users` WHERE `userId` = :adminId";
+$cmdCheck = $db->prepare($sqlCheck);
+$cmdCheck->bindParam("adminId", $adminId, PDO::PARAM_INT);
+$cmdCheck->execute();
+$arrCheck = $cmdCheck->fetch();
+if (!empty($arrCheck)) {
+  $sql = "DELETE FROM `cms_users` WHERE `userId` = :adminId";
+  $cmd = $db->prepare($sql);
+  $cmd->bindParam(":adminId", $adminId, PDO::PARAM_INT);
+  $cmd->execute();
+  header("location:admins.php");
+  //Deletes the entry with the assocciated id, link back to admins
+} else {
+  echo "<p>An error occurred. The admin you wish to delete doesn't exist.</p>";
+}
 $db = null;
-header("location:admins.php");
-//Links back to admins
+//Close database connection
+require "footer.php";
+//Header and footer are displayed if an error message keeps the user from continuing
